Set second argument of split to val
According to the "Supported Functions" section of the
docs (https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-split.html),
the second argument to Split doesn't have to be a string literal and
can instead be a function. The delimiter must be a string literal.

Fun note: this wackiness is needed to be able to prepend/append values
to a List type parameter.