Remove support for custom (enum) type overrides
This is a good feature to have. And I'd love to bring it back in a more
maintainable setting. But for now it has to go to spend my limited time
budget on higher ROI tasks.

The full story:

The cloud formation specification has many `String` fields where the values are
not free text but only certain values are valid.

Valid values are not exposed in the cloud formation spec, and are only
available in the documentation rich text.

For this reason stratosphere used to ship a manual maintained set of overwrites that
would map resource properties to manually defined enum-a-like types.

This manually maintained set of enum types often goes stale and leads to issues
where AWS would release a new supported value and the stratosphere library was not
yet updated to support that new value - as a simple regeneration from the spec is not enough
to catch these.

This feature can be brought back with this thoughts in mind:

* We may be able to parse the enum values from parsing the documentation string.
* If we bring it back each enum type needs to have a "custom" type member so library
  users always have an escape hatch for stratosphere lagging behind AWS.